+++ src/api/todolists-api.ts
@@ -1,4 +1,11 @@
 import axios from 'axios'
+import {
+   GetTasksResponse,
+   TodolistType,
+   UpdateTaskModelType,
+   ResponseType,
+} from '../common/types/types'
+import { LoginParamsType, TaskType } from '../common/types/types'
 
 const settings = {
    withCredentials: true,
@@ -41,13 +48,6 @@ export const todolistsAPI = {
    },
 }
 
-export type LoginParamsType = {
-   email: string
-   password: string
-   rememberMe: boolean
-   captcha?: string
-}
-
 export const authAPI = {
    login(data: LoginParamsType) {
       return instance.post<ResponseType<{ userId?: number }>>('auth/login', data)
@@ -58,55 +58,4 @@ export const authAPI = {
    me() {
       return instance.get<ResponseType<{ id: number; email: string; login: string }>>('auth/me')
    },
-}
-
-// types
-export type TodolistType = {
-   id: string
-   title: string
-   addedDate: string
-   order: number
-}
-export type ResponseType<D = {}> = {
-   resultCode: number
-   messages: Array<string>
-   data: D
-}
-export enum TaskStatuses {
-   New = 0,
-   InProgress = 1,
-   Completed = 2,
-   Draft = 3,
-}
-export enum TaskPriorities {
-   Low = 0,
-   Middle = 1,
-   Hi = 2,
-   Urgently = 3,
-   Later = 4,
-}
-export type TaskType = {
-   description: string
-   title: string
-   status: TaskStatuses
-   priority: TaskPriorities
-   startDate: string
-   deadline: string
-   id: string
-   todoListId: string
-   order: number
-   addedDate: string
-}
-export type UpdateTaskModelType = {
-   title: string
-   description: string
-   status: TaskStatuses
-   priority: TaskPriorities
-   startDate: string
-   deadline: string
-}
-type GetTasksResponse = {
-   error: string | null
-   totalCount: number
-   items: TaskType[]
 }
