+++ src/features/TodolistsList/TodolistsList.tsx
@@ -1,17 +1,14 @@
 import React, { useCallback, useEffect } from 'react'
 import { Grid, Paper } from '@material-ui/core'
-import { AddItemForm } from '../../components/AddItemForm/AddItemForm'
+import { AddItemForm } from '../../common/components/AddItemForm/AddItemForm'
 import { Todolist } from './Todolist/Todolist'
 import { Redirect } from 'react-router-dom'
 import { observer } from 'mobx-react-lite'
 import auth from '../../store/auth'
 import todolists from '../../store/todolists'
+import { ComponentType } from '../../common/types/types'
 
-type PropsType = {
-   demo?: boolean
-}
-
-export const TodolistsList: React.FC<PropsType> = observer(({ demo }) => {
+export const TodolistsList: React.FC<ComponentType> = observer(({ demo }) => {
    const todolistsItems = todolists.todos
    const isLoggedIn = auth.isLoggedIn
 
@@ -30,18 +27,16 @@ export const TodolistsList: React.FC<PropsType> = observer(({ demo }) => {
    return (
       <>
          <Grid container style={{ padding: '20px' }}>
-            <AddItemForm addItem={addTodolist} />
+            <AddItemForm addItem={addTodolist} placeholder={'new List'} />
          </Grid>
-         <Grid container spacing={3}>
-            {todolistsItems.map(tl => {
-               return (
-                  <Grid item key={tl.id}>
-                     <Paper style={{ padding: '10px' }}>
-                        <Todolist todolist={tl} demo={demo} />
-                     </Paper>
-                  </Grid>
-               )
-            })}
+         <Grid container spacing={3} justifyContent={'center'}>
+            {todolistsItems.map(tl => (
+               <Grid item key={tl.id}>
+                  <Paper style={{ padding: '10px' }}>
+                     <Todolist todolist={tl} demo={demo} />
+                  </Paper>
+               </Grid>
+            ))}
          </Grid>
       </>
    )
